backend/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
openai
python-dotenv
beanie
motor
